Moved middleware to utils directory.
